+++ gnbcu/src/datastore/ue_state_store.rs
@@ -4,7 +4,7 @@ use async_trait::async_trait;
 
 #[async_trait]
 pub trait UeStateStore: Clone + Send + Sync + 'static {
-    async fn store(&self, k: u32, s: UeState, ttl_secs: u32) -> Result<()>;
+    async fn store(&self, k: u32, s: UeState, ttl_secs: usize) -> Result<()>;
     async fn retrieve(&self, k: &u32) -> Result<Option<UeState>>;
     async fn delete(&self, k: &u32) -> Result<()>;
 }
