+++ gnbcu/src/lib.rs
@@ -124,7 +124,7 @@ impl<U: UeStateStore> Gnbcu<U> {
 
 #[async_trait]
 impl<U: UeStateStore> UeStateStore for Gnbcu<U> {
-    async fn store(&self, k: u32, s: UeState, ttl_secs: u32) -> Result<()> {
+    async fn store(&self, k: u32, s: UeState, ttl_secs: usize) -> Result<()> {
         self.ue_store.store(k, s, ttl_secs).await
     }
     async fn retrieve(&self, k: &u32) -> Result<Option<UeState>> {
